Prototype iframe idea
Prototype idea for using iframes for page content.  Currently, there is
an array that contains a list of iframe sources and upon page loading,
the JS creates all iframes.  Calendar, Syllabus and the “Home Page” are
only functional at this time.  default.html is the file that can
execute this prototype